refactor: apply reviewed ruff 0.16 unsafe autofixes
75 fixes, applied and read rule by rule rather than in one blanket pass:

- RUF046 (32): drop the redundant int() around single-arg round(), which
  already returns int (verified on numpy scalars too, not just floats).
- ISC004 (9): parenthesize implicit string concatenation inside list literals.
  All nine were intentional line wraps, not missing commas.
- C408 dict() -> {}, RUF059 unused unpacked vars -> _, PIE810 startswith tuples,
  RUF007 zip(x, x[1:]) -> itertools.pairwise, SIM118 .keys() drops, UP031/UP032
  %-format -> f-string, UP022 capture_output=True, plus assorted singles.

Hand-cleaned two autofix artifacts: a redundant bool() wrapper in
ollama_client._is_connection_refused and redundant parens in titlecards.

106 -> 33 findings. Full suite green (2238 passed).

+++ screenspace_frames.py
@@ -189,8 +189,8 @@ def _ffmpeg_pipe_frames(
     # max_dim cap. Skipped at 1.0 so default-config runs are byte-identical
     # to pre-feature behavior.
     if cv_scale > 0 and abs(cv_scale - 1.0) > 1e-6:
-        scaled_w = max(2, int(round(out_w * cv_scale)))
-        scaled_h = max(2, int(round(out_h * cv_scale)))
+        scaled_w = max(2, round(out_w * cv_scale))
+        scaled_h = max(2, round(out_h * cv_scale))
         scaled_w += scaled_w % 2
         scaled_h += scaled_h % 2
         filters.append(f"scale={scaled_w}:{scaled_h}:flags=lanczos")
